requirements.txt
openai
python-dotenv
googlemaps
fastapi
uvicorn